+++ docs/source/symbac_drawing.rst
@@ -1,0 +1,5 @@
+SyMBac.drawing
+==========================
+
+.. automodule:: SyMBac.drawing
+    :members:
