--- docs/source/symbac_general_drawing.rst
@@ -1,5 +1,0 @@
-SyMBac.general_drawing
-==========================
-
-.. automodule:: SyMBac.general_drawing
-    :members:
