--- docs/source/symbac_phase_contrast_drawing.rst
@@ -1,5 +1,0 @@
-SyMBac.phase_contrast_drawing
-================================
-
-.. automodule:: SyMBac.phase_contrast_drawing
-    :members:
