+++ docs/source/symbac_pySHINE.rst
@@ -1,0 +1,5 @@
+SyMBac.pySHINE
+==========================
+
+.. automodule:: SyMBac.pySHINE
+    :members:
